Move `DeployAWS` into the `AWS` module
This is the final file that needs to be in the `AWS` module. This will
most likely be renamed a helper module
